Rewrited the update methods by removing the fields that doesn't necessarily need to be removed at this step and handled "NOT FOUND" exception for DELETE methods.

+++ src/main/java/com/codefactorygroup/betting/service/implementations/CompetitionServiceImpl.java
@@ -2,19 +2,19 @@ package com.codefactorygroup.betting.service.implementations;
 
 import com.codefactorygroup.betting.converter.CompetitionDTOtoCompetitionConverter;
 import com.codefactorygroup.betting.converter.EventDTOtoEventConverter;
-import com.codefactorygroup.betting.converter.SportDTOtoSportConverter;
 import com.codefactorygroup.betting.domain.Competition;
 import com.codefactorygroup.betting.domain.Event;
+import com.codefactorygroup.betting.domain.Selection;
 import com.codefactorygroup.betting.dto.CompetitionDTO;
 import com.codefactorygroup.betting.dto.EventDTO;
 import com.codefactorygroup.betting.exception.EntityAlreadyExistsException;
 import com.codefactorygroup.betting.exception.NoSuchEntityExistsException;
 import com.codefactorygroup.betting.repository.CompetitionRepository;
 import com.codefactorygroup.betting.service.CompetitionService;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -26,13 +26,10 @@ public class CompetitionServiceImpl implements CompetitionService {
     private final CompetitionDTOtoCompetitionConverter competitionDTOtoCompetitionConverter;
     private final EventDTOtoEventConverter eventDTOtoEventConverter;
 
-    private final SportDTOtoSportConverter sportDTOtoSportConverter;
-
-    public CompetitionServiceImpl(CompetitionRepository competitionRepository, CompetitionDTOtoCompetitionConverter competitionDTOtoCompetitionConverter, EventDTOtoEventConverter eventDTOtoEventConverter, SportDTOtoSportConverter sportDTOtoSportConverter) {
+    public CompetitionServiceImpl(CompetitionRepository competitionRepository, CompetitionDTOtoCompetitionConverter competitionDTOtoCompetitionConverter, EventDTOtoEventConverter eventDTOtoEventConverter) {
         this.competitionRepository = competitionRepository;
         this.competitionDTOtoCompetitionConverter = competitionDTOtoCompetitionConverter;
         this.eventDTOtoEventConverter = eventDTOtoEventConverter;
-        this.sportDTOtoSportConverter = sportDTOtoSportConverter;
     }
 
     @Override
@@ -76,7 +73,7 @@ public class CompetitionServiceImpl implements CompetitionService {
                 .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Competition with ID=%d doesn't exist.", competitionId)));
 
         Event event = eventDTOtoEventConverter.convert(eventDTO);
-        competition.getEvents().add(event);
+        competition.addEvent(event);
 
         return CompetitionDTO.converter(competitionRepository.save(competition));
     }
@@ -84,18 +81,16 @@ public class CompetitionServiceImpl implements CompetitionService {
     @Transactional
     @Override
     public void deleteCompetition(Integer competitionId) {
-        competitionRepository.deleteById(competitionId);
+        boolean competitionExists = competitionRepository.existsById(competitionId);
+        if (competitionExists) {
+            competitionRepository.deleteById(competitionId);
+        } else {
+            throw new NoSuchEntityExistsException(String.format("Competition with ID=%d doesn't exist.", competitionId));
+        }
     }
 
-
     private Competition update(final Competition competition, final CompetitionDTO toUpdateCompetition) {
-        List<EventDTO> eventDTOS = Optional.of(toUpdateCompetition).map(CompetitionDTO::events).orElseGet(Collections::emptyList);
-        competition.setId(toUpdateCompetition.id());
         competition.setName(toUpdateCompetition.name());
-        competition.setEvents(eventDTOS
-                .stream()
-                .map(eventDTOtoEventConverter::convert)
-                .collect(Collectors.toList()));
         return competition;
     }
