+++ src/main/java/com/codefactorygroup/betting/service/implementations/EventServiceImpl.java
@@ -3,11 +3,9 @@ package com.codefactorygroup.betting.service.implementations;
 import com.codefactorygroup.betting.converter.EventDTOtoEventConverter;
 import com.codefactorygroup.betting.converter.MarketDTOtoMarketConverter;
 import com.codefactorygroup.betting.converter.ParticipantDtoToParticipantConverter;
-import com.codefactorygroup.betting.domain.Competition;
 import com.codefactorygroup.betting.domain.Event;
 import com.codefactorygroup.betting.domain.Market;
 import com.codefactorygroup.betting.domain.Participant;
-import com.codefactorygroup.betting.dto.CompetitionDTO;
 import com.codefactorygroup.betting.dto.EventDTO;
 import com.codefactorygroup.betting.dto.MarketDTO;
 import com.codefactorygroup.betting.dto.ParticipantDTO;
@@ -15,6 +13,7 @@ import com.codefactorygroup.betting.exception.EntityAlreadyExistsException;
 import com.codefactorygroup.betting.exception.NoSuchEntityExistsException;
 import com.codefactorygroup.betting.repository.EventRepository;
 import com.codefactorygroup.betting.service.EventService;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -99,7 +98,7 @@ public class EventServiceImpl implements EventService {
                 .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Event with ID=%d doesn't exist.", eventId)));
 
         Market market = marketDTOtoMarketConverter.convert(marketDTO);
-        event.getMarkets().add(market);
+        event.addMarket(market);
 
         return EventDTO.converter(eventRepository.save(event));
     }
@@ -111,35 +110,27 @@ public class EventServiceImpl implements EventService {
                 .orElseThrow(() -> new NoSuchEntityExistsException(String.format("Event with ID=%d doesn't exist.", eventId)));
 
         Participant participant = participantDtoToParticipantConverter.convert(participantDTO);
-        event.getParticipants().add(participant);
+        event.addParticipant(participant);
 
         return EventDTO.converter(eventRepository.save(event));
     }
 
-
-
     @Transactional
     @Override
     public void deleteEvent(Integer eventId) {
-        eventRepository.deleteById(eventId);
+        boolean eventExists = eventRepository.existsById(eventId);
+        if (eventExists) {
+            eventRepository.deleteById(eventId);
+        } else {
+            throw new NoSuchEntityExistsException(String.format("Event with ID=%d doesn't exist.", eventId));
+        }
     }
 
 
     private Event update(final Event event, final EventDTO toUpdateEvent) {
-        List <ParticipantDTO> participantDTOS = Optional.of(toUpdateEvent).map(EventDTO::participants).orElseGet(Collections::emptyList);
-        List <MarketDTO> marketDTOS = Optional.of(toUpdateEvent).map(EventDTO::markets).orElseGet(Collections::emptyList);
-        event.setId(toUpdateEvent.id());
         event.setName(toUpdateEvent.name());
         event.setStartTime(toUpdateEvent.startTime());
         event.setEndTime(toUpdateEvent.endTime());
-        event.setParticipants(participantDTOS
-                .stream()
-                .map(participantDtoToParticipantConverter::convert)
-                .collect(Collectors.toList()));
-        event.setMarkets(marketDTOS
-                .stream()
-                .map(marketDTOtoMarketConverter::convert)
-                .collect(Collectors.toList()));
         return event;
     }
